Split Control.Monad in two, and add Data.<$
