+++ src/Prelude/Linear.hs
@@ -89,10 +89,6 @@ forget f x = f x
 flip :: (a -->.(p) b -->.(q) c) -->.(r) b -->.(q) a -->.(p) c
 flip f b a = f a b
 
--- | Linearly typed replacement for the standard '(Prelude.<$)' function.
-(<$) :: (Control.Functor f, Consumable b) => a ->. f b ->. f a
-a <$ fb = Control.fmap (`lseq` a) fb
-
 -- | Linearly typed replacement for the standard '(Prelude.<*)' function.
 (<*) :: (Data.Applicative f, Consumable b) => f a ->. f b ->. f a
 fa <* fb = Data.fmap (flip lseq) fa Data.<*> fb
